requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-bootstrap4==23.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
idna==3.4
mysql-connector-python==8.0.28
Pillow==9.5.0
pytz
requests==2.28.2
soupsieve==2.4.1
sqlparse==0.4.3
urllib3==1.26.15
whitenoise==5.2.0
